% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_spotoroo.R
\name{dynamic_spotoroo}
\alias{dynamic_spotoroo}
\title{Plotting spatiotemporal clustering results spatially: with fire movement}
\usage{
dynamic_spotoroo(
  results,
  clusters = "all",
  hotspots = TRUE,
  noises = FALSE,
  bottom = NULL
)
}
\arguments{
\item{results}{an object of class "\code{spotoroo}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{clusters}{character/numeric; if "all", plot all clusters. if a numeric
vector is given, plot corresponding clusters. plotting more
than one cluster is not recommended by this function.}

\item{hotspots}{logical; if \code{TRUE}, plot the hotspots.}

\item{noises}{logical; if \code{TRUE}, plot the noises. plotting noises is
not recommended by this function.}

\item{bottom}{an object of class "\code{ggplot}", optional; if \code{TRUE},
plot onto this object. Now it only supports object without colour related
aesthetics.}
}
\value{
an object of class "\code{ggplot}".
}
\description{
can be called by \code{plot.spotoroo()} or \code{plot_spotoroo()}.
}
\examples{
data("hotspots500")
results <- hotspot_cluster(hotspots500,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)

dynamic_spotoroo(results, clusters = 1)
}
