% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fire.R
\name{extract_fire}
\alias{extract_fire}
\title{Extracting fires from the spatiotemporal clustering results}
\usage{
extract_fire(result, cluster = "all", noise = FALSE)
}
\arguments{
\item{result}{\code{spotoroo} object;
a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{cluster}{character/integer; if "all", extract all clusters.
if a integer vector is given, extract corresponding
clusters.}

\item{noise}{logical; whether or not to include noise.}
}
\description{
\code{extract_fire()} takes a \code{spotoroo} object to produce a data frame which
contains information about the fire.
}
\examples{
result <- hotspot_cluster(hotspots_fin,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)



a <- extract_fire(result)
a[1:10, ]


a <- extract_fire(result, 4)
a[1:10, ]


}
