% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fire_mov.R
\name{get_fire_mov}
\alias{get_fire_mov}
\title{Calculate fire movement}
\usage{
get_fire_mov(result, cluster, step = 1, method = "mean")
}
\arguments{
\item{result}{spotoroo result object; get from \code{\link{hotspot_cluster}}.}

\item{cluster}{integer; the membership lable of the fire.}

\item{step}{integer; step size.}

\item{method}{character; "mean" or "median", method of centre calculation.}
}
\value{
a data frame of fire movement
\itemize{
\item \code{membership} : membership labels.
\item \code{lon} : longitude of fire centre.
\item \code{lat} : latitude of fire centre.
\item \code{timeID} : time indexes.
\item \code{obsTime} : observed time (approximated).
\item \code{ignition} : whether or not it is a ignition point.
}
}
\description{
Calculate the movement of a single fire per a given timestamps.
}
\examples{
result <- hotspot_cluster(hotspots_fin,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

get_fire_mov(result, cluster = 1, step = 1, method = "mean")
get_fire_mov(result, cluster = 2, step = 3, method = "median")

}
