% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_noises.R
\name{handle_noises}
\alias{handle_noises}
\title{Handle noise clusters}
\usage{
handle_noises(global_memberships, minPts)
}
\arguments{
\item{global_memberships}{integer; a vector of membership labels.}

\item{minPts}{numeric (>=0); minimum number of points of a cluster.}
}
\value{
integer; a vector of membership labels.
}
\description{
Handle noise clusters with a given threshold (minimum number of points of
a cluster). Noises will be assigned with membership label \code{-1}. Normal
clusters will be assigned with adjusted membership labels.
}
\examples{
global_memberships <- c(1,1,1,2,2,2,2,2,2,3,3,3,3,3,3)

handle_noises(global_memberships, 4)

# -1 -1 -1 1 1 1 1 1 1 2 2 2 2 2 2
}
