% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspots}
\alias{hotspots}
\title{1070 observations of satellite hot spots}
\format{
A data frame with 1070 rows and 3 variables:
\describe{
\item{lon}{longitude}
\item{lat}{latitude}
\item{obsTime}{observed time}
}
}
\source{
\url{https://www.eorc.jaxa.jp/ptree/}
}
\usage{
hotspots
}
\description{
A dataset containing the 1070 observations of satellite hot spots in
Victoria, Australia during the 2019-2020 Australian bushfire season.
}
\keyword{datasets}
