% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspots500}
\alias{hotspots500}
\title{500 observations of satellite hotspots.}
\format{
A data frame with 500 rows and 4 variables:
\describe{
\item{id}{unique identifier}
\item{lon}{longitude}
\item{lat}{latitude}
\item{obsTime}{observed time}
}
}
\source{
\url{https://www.eorc.jaxa.jp/ptree/}
}
\usage{
hotspots500
}
\description{
A dataset containing the 500 observations of satellite hotspots in Victoria,
Australia during the 2019-2020 Australian bushfire season.
}
\keyword{datasets}
