% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignition_points.R
\name{ignition_points}
\alias{ignition_points}
\title{Calculate ignition points}
\usage{
ignition_points(lon, lat, obsTime, timeID, memberships, ignitionCenter)
}
\arguments{
\item{lon}{numeric; a vector of longitude values.}

\item{lat}{numeric; a vector of latitude values.}

\item{obsTime}{date/numeric; a vector of observed time.}

\item{timeID}{integer (>=1); a vector of time indexes.}

\item{memberships}{integer; a vector of membership labels.}

\item{ignitionCenter}{character; method of calculating ignition points,
one of "mean" and "median".}
}
\value{
a data frame of ignition points
\itemize{
\item \code{ignition_lon} : longitude of ignition points.
\item \code{ignition_lat} : latitude of ignition points.
\item \code{ignition_obsTime} : observed time of ignition points.
}
}
\description{
Calculate ignition points using membership labels.
}
\details{
For a cluster, when there are multiple earliest hotspots, if
ignitionCenter = "mean", the centroid of these hotspots will be used
as the ignition point, if ignitionCenter = "median", median longitude
and median latitude of these hotspots will be used.
}
\examples{
lon <- c(141.1, 141.14, 141.12, 141.14, 141.16, 141.12, 141.14,
          141.16, 141.12, 141.14)
lat <- c(-37.10, -37.10, -37.12, -37.12, -37.12, -37.14, -37.14,
         -37.14, -37.16, -37.16)
obsTime <- c(rep(1, 5), rep(26, 5))
memberships <- c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2)

global_clustering(lon, lat, obsTime, memberships, "mean")

#   ignition_lon ignition_lat ignition_obsTime
# 1      141.132      -37.112                1
# 2      141.136      -37.148               26
}
