% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.hotspotcluster}
\alias{plot.hotspotcluster}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{hotspotcluster}(results, draw_ignitions = TRUE, draw_hotspots = TRUE, bottom = NULL, ...)
}
\arguments{
\item{results}{an object of class "\code{hotspotcluster}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{draw_ignitions}{logical; if \code{TRUE}, plot the ignition points.}

\item{draw_hotspots}{logical; if \code{TRUE}, plot the hotspots.}

\item{bottom}{an object of class "\code{ggplot}"", optional; if \code{TRUE},
plot onto this object.}

\item{...}{further arguments will be omitted.}
}
\value{
an object of class "\code{ggplot}",
}
\description{
\code{plot} method for class "\code{hotspotcluster}"
}
\examples{
results <- hotspot_cluster(hotspots5000,
                           lon = "lon",
                           lat = "lat",
                           obstime = "obstime",
                           ActiveTime = 24,
                           AdjDist = 3000,
                           ignition_center = "mean",
                           time_unit = "hours",
                           timestep = 1)
plot(results)
}
