% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.spotoroo}
\alias{plot.spotoroo}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{spotoroo}(x, ...)
}
\arguments{
\item{x}{\code{spotoroo} object;
a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{\dots}{additional arguments pass to \code{\link[=plot_spotoroo]{plot_spotoroo()}}}
}
\description{
\code{\link[=plot.spotoroo]{plot.spotoroo()}} is the \code{plot} method of the class \code{spotoroo}.
It is a simple wrapper of \code{\link[=plot_spotoroo]{plot_spotoroo()}}.
}
\examples{
results <- hotspot_cluster(hotspots_fin,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           minTime = 3,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)

# default plot
plot(results)

# three different types of plots
plot(results, "def")
plot(results, "mov")
plot(results, "timeline")

}
