% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot_results}
\alias{plot_results}
\title{Plotting spatiotemporal clustering results}
\usage{
plot_results(
  results,
  drawIgnitions = TRUE,
  drawHotspots = TRUE,
  drawNoises = TRUE,
  bottom = NULL
)
}
\arguments{
\item{results}{an object of class "\code{spotoroo}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{drawIgnitions}{logical; if \code{TRUE}, plot the ignition points.}

\item{drawHotspots}{logical; if \code{TRUE}, plot the hotspots.}

\item{drawNoises}{sldfjks}

\item{bottom}{an object of class "\code{ggplot}", optional; if \code{TRUE},
plot onto this object.}
}
\value{
an object of class "\code{ggplot}".
}
\description{
can be called by \code{plot.spotoroo()}.
}
\examples{
data("hotspots500")
results <- hotspot_cluster(hotspots500,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)
plot(results)
}
