% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vic_map.R
\name{plot_vic_map}
\alias{plot_vic_map}
\title{Plotting map of Victoria, Australia}
\usage{
plot_vic_map(...)
}
\arguments{
\item{...}{all arguments will be ignored.}
}
\value{
\code{ggplot} object; the map of Victoria, Australia.
}
\description{
This function plots the map of Victoria, Australia.
}
\details{
Require package \code{sf} installed.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  plot_vic_map()
}

}
