% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotoroo-package.R
\docType{package}
\name{spotoroo-package}
\alias{spotoroo}
\alias{spotoroo-package}
\title{spotoroo: Spatiotemporal Clustering of Satellite Hot Spot Data}
\description{
An algorithm to cluster satellite hot spot data spatially and temporally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tengmcing.github.io/spotoroo/}
  \item \url{https://github.com/TengMCing/spotoroo/}
  \item Report bugs at \url{https://github.com/TengMCing/spotoroo/issues}
}

}
\author{
\strong{Maintainer}: Weihao Li \email{llreczx@gmail.com} (\href{https://orcid.org/0000-0003-4959-106X}{ORCID})

Other contributors:
\itemize{
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID}) [contributor]
  \item Emily Dodwell \email{emdodwell@gmail.com} [contributor]
}

}
\keyword{internal}
