% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\docType{package}
\name{spotoroo}
\alias{spotoroo}
\title{spotoroo: spatiotemporal clustering in R of hot spot}
\description{
A package for clustering satellite hot spots and detecting
fire ignition points.
}
\section{Authors}{

\itemize{
\item{Weihao Li \url{weihao.li@monash.edu}}
\item{Dianne Cook \url{dicook@monash.edu}}
\item{Emily Dodwell \url{emdodwell@gmail.com}}
}
}

\section{Functions}{

The spotoroo package provides 13 important functions:
\itemize{
\item{\code{\link[=hotspot_cluster]{hotspot_cluster()}}}
\item{\code{\link[=global_clustering]{global_clustering()}}}
\item{\code{\link[=local_clustering]{local_clustering()}}}
\item{\code{\link[=handle_noise]{handle_noise()}}}
\item{\code{\link[=ignition_point]{ignition_point()}}}
\item{\code{\link[=get_fire_mov]{get_fire_mov()}}}
\item{\code{\link[=plot.spotoroo]{plot.spotoroo()}}}
\item{\code{\link[=plot_spotoroo]{plot_spotoroo()}}}
\item{\code{\link[=plot_def]{plot_def()}}}
\item{\code{\link[=plot_fire_mov]{plot_fire_mov()}}}
\item{\code{\link[=plot_timeline]{plot_timeline()}}}
\item{\code{\link[=plot_vic_map]{plot_vic_map()}}}
\item{\code{\link[=summary.spotoroo]{summary.spotoroo()}}}
\item{\code{\link[=summary_spotoroo]{summary_spotoroo()}}}
\item{\code{\link[=print.spotoroo]{print.spotoroo()}}}
\item{\code{\link[=transform_time_id]{transform_time_id()}}}
\item{\code{\link[=dist_point_to_vector]{dist_point_to_vector()}}}
}

The spotoroo package provides 2 external data objects:
\itemize{
\item{\link{hotspots}}
\item{\link{vic_map}}
}
}

