% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{summary.spotoroo}
\alias{summary.spotoroo}
\title{Summarizing spatiotemporal clustering result}
\usage{
\method{summary}{spotoroo}(object, ...)
}
\arguments{
\item{object}{\code{spotoroo} object;
a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{\dots}{additional arguments pass to \code{\link[=summary_spotoroo]{summary_spotoroo()}}}
}
\description{
\code{summary.spotoroo()} is the \code{summary} method of the class \code{spotoroo}.
It is a simple wrapper of \code{\link[=summary_spotoroo]{summary_spotoroo()}}.
}
\examples{
result <- hotspot_cluster(hotspots_fin,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           minTime = 3,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)



summary(result)


}
