% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_time_id.R
\name{transform_time_id}
\alias{transform_time_id}
\title{Transform observed time to time indexes}
\usage{
transform_time_id(obsTime, timeUnit, timeStep)
}
\arguments{
\item{obsTime}{date/numeric; a vector of observed time of hotspots. If
\code{timeUnit} = \code{"n"}, \code{obsTime}
can be a numeric vector, otherwise, it needs
to be in date time format.}

\item{timeUnit}{character; the unit of time, one of "s" (secs), "m"(mins),
"h"(hours), "d"(days) and "n"(numeric).}

\item{timeStep}{numeric (>0); the number of units of timeUnit as a time
step.}
}
\value{
integer; a vector of time indexes.
}
\description{
Transform observed time to time indexes.
}
\details{
The earliest observed time has a corresponding time index \code{1}. The
difference between any time to the earliest observed time will be transformed
using the given time unit and divided by the given time step. The final
indexes will be floored to integer.
}
\examples{
obsTime <- as.Date(c("2020-01-01",
                     "2020-01-02",
                     "2020-01-04"))

transform_time_id(obsTime, "h", 1)

# 1 25 73
}
