% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_point_to_vector.R
\name{dist_point_to_vector}
\alias{dist_point_to_vector}
\title{Calculate the geodesic of a point to multiple points}
\usage{
dist_point_to_vector(plon, plat, vlon, vlat)
}
\arguments{
\item{plon}{numeric; the longitude of a point.}

\item{plat}{numeric; the latitude of a point.}

\item{vlon}{numeric; a vector of longitude values.}

\item{vlat}{numeric; a vector of latitude values.}
}
\value{
numeric; the geodesic of a point to multiple points in meters.
}
\description{
Calculate the geodesic of a point to multiple points given coordinate
information.
The calculation of the geodesic uses the mapbox cheap ruler.
See also \code{\link[geodist]{geodist}} for more details.
}
\examples{
vlon <- c(141.12, 141.13)
vlat <- c(-37.1, -37.0)

dist_point_to_vector(141.12, -37.1, vlon, vlat)

# 0.00 11148.63
}
