% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fire.R
\name{extract_fire}
\alias{extract_fire}
\title{Extracting fires from the spatiotemporal clustering results}
\usage{
extract_fire(result, cluster = "all", noise = FALSE)
}
\arguments{
\item{result}{\code{spotoroo} object;
a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{cluster}{character/integer; if "all", extract all clusters.
if a integer vector is given, extract corresponding
clusters.}

\item{noise}{logical; whether or not to include noise.}
}
\value{
data.frame; the fire information
\itemize{
\item \code{lon} : longitude.
\item \code{lat} : latitude.
\item \code{obsTime} : observed time.
\item \code{timeID} : time indexes.
\item \code{membership} : membership labels.
\item \code{noise} : whether it is a noise point.
\item \code{distToIgnition} : distance to the ignition location.
\item \code{distToIgnitionUnit} : unit of distance to the ignition
location.
\item \code{timeFromIgnition} : time from ignition.
\item \code{timeFromIgnitionUnit} : unit of time from ignition.
\item \code{type} : type of the entry, either "hotspot", "noise" or
"ignition"
\item \code{obsInCluster} : number of observations in the cluster.
\item \code{clusterTimeLen} : length of time of the cluster.
\item \code{clusterTimeLenUnit} : unit of length of time of the
cluster.
}
}
\description{
This function takes a \code{spotoroo} object to produce a data frame which
contains information about the fire.
}
\examples{
# get clustering results
result <- hotspot_cluster(hotspots,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)


# extract all fires
all_fires <- extract_fire(result)
head(all_fires, 3)


fire_4 <- extract_fire(result, 4)
head(fire_4, 3)


}
