% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_clustering.R
\name{global_clustering}
\alias{global_clustering}
\title{Cluster hotspots spatially and temporally}
\usage{
global_clustering(lon, lat, timeID, activeTime, adjDist)
}
\arguments{
\item{lon}{numeric; a vector of longitude values.}

\item{lat}{numeric; a vector of latitude values.}

\item{timeID}{integer (>=1); a vector of time indexes.}

\item{activeTime}{numeric; time tolerance.
see also \code{\link{hotspot_cluster}}.}

\item{adjDist}{numeric; distance tolerance.
see also \code{\link{hotspot_cluster}}.}
}
\value{
integer; a vector of membership labels.
}
\description{
Cluster hotspots spatially and temporally.
}
\examples{
lon <- c(141.1, 141.14, 141.12, 141.14, 141.16, 141.12, 141.14,
          141.16, 141.12, 141.14)
lat <- c(-37.10, -37.10, -37.12, -37.12, -37.12, -37.14, -37.14,
         -37.14, -37.16, -37.16)
timeID <- c(rep(1, 5), rep(26, 5))

global_clustering(lon, lat, timeID, 24, 3000)

# 1 1 1 1 1 2 2 2 2 2
}
