% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_noise.R
\name{handle_noise}
\alias{handle_noise}
\title{Handle noise clusters}
\usage{
handle_noise(global_membership, minPts)
}
\arguments{
\item{global_membership}{integer; a vector of membership labels.}

\item{minPts}{numeric (>=0); minimum number of points of a cluster.}
}
\value{
integer; a vector of membership labels.
}
\description{
Handle noise clusters with a given threshold (minimum number of points of
a cluster). noise will be assigned with membership label \code{-1}. Normal
clusters will be assigned with adjusted membership labels.
}
\examples{
global_membership <- c(1,1,1,2,2,2,2,2,2,3,3,3,3,3,3)

handle_noise(global_membership, 4)

# -1 -1 -1 1 1 1 1 1 1 2 2 2 2 2 2
}
