% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{hotspot_cluster}
\alias{hotspot_cluster}
\title{Spatiotemporal clustering of hotspots}
\usage{
hotspot_cluster(
  hotspots,
  lon = "lon",
  lat = "lat",
  obsTime = "obsTime",
  activeTime = 24,
  adjDist = 3000,
  minPts = 4,
  ignitionCenter = "mean",
  timeUnit = NULL,
  timeStep = NULL
)
}
\arguments{
\item{hotspots}{list/data frame; a list contains information of hotspots.}

\item{lon}{character; the name of the column of the list which contains
longitude values.}

\item{lat}{character; the name of the column of the list which contains
latitude values.}

\item{obsTime}{character; the name of the column of the list which contains
observed time.}

\item{activeTime}{numeric (>=0); time tolerance.}

\item{adjDist}{numeric (>0); distance tolerance.}

\item{minPts}{numeric (>0); minimum number of points in a cluster.}

\item{ignitionCenter}{character;}

\item{timeUnit}{sldfjks}

\item{timeStep}{sldfjks}
}
\value{
integer; a vector of membership labels.
}
\description{
Cluster hotspots spatially and temporally.
}
\examples{
global_memberships <- c(1,1,1,2,2,2,2,2,2,3,3,3,3,3,3)

handle_noises(global_memberships, 4)

# -1 -1 -1 1 1 1 1 1 1 2 2 2 2 2 2
}
