% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspots5000}
\alias{hotspots5000}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{id}{price, in US dollars}
\item{lat}{weight of the diamond, in carats}
\item{lon}{price, in US dollars}
\item{obsTime}{weight of the diamond, in carats}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
hotspots5000
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
