% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignition_point.R
\name{ignition_point}
\alias{ignition_point}
\title{Calculate ignition points}
\usage{
ignition_point(lon, lat, obsTime, timeUnit, timeID, membership, ignitionCenter)
}
\arguments{
\item{lon}{numeric; a vector of longitude values.}

\item{lat}{numeric; a vector of latitude values.}

\item{obsTime}{date/numeric; a vector of observed time.}

\item{timeUnit}{character; the unit of time, one of "s" (secs), "m"(mins),
"h"(hours), "d"(days) and "n"(numeric).}

\item{timeID}{integer (>=1); a vector of time indexes.}

\item{membership}{integer; a vector of membership labels.}

\item{ignitionCenter}{character; method of calculating ignition points,
one of "mean" and "median".}
}
\value{
a data frame of ignition points
\itemize{
\item \code{membership} : membership labels.
\item \code{ignition_lon} : longitude of ignition points.
\item \code{ignition_lat} : latitude of ignition points.
\item \code{ignition_obsTime} : observed time of ignition points.
\item \code{timeID} : time indexes.
\item \code{clusterObs} : number of observations in the cluster.
\item \code{clusterTimeLen} : time frame of the cluster.
}
}
\description{
Calculate ignition points using membership labels.
}
\details{
For a cluster, when there are multiple earliest hotspots, if
ignitionCenter = "mean", the centroid of these hotspots will be used
as the ignition point, if ignitionCenter = "median", median longitude
and median latitude of these hotspots will be used.
}
\examples{
lon <- c(141.1, 141.14, 141.12, 141.14, 141.16, 141.12, 141.14,
          141.16, 141.12, 141.14)
lat <- c(-37.10, -37.10, -37.12, -37.12, -37.12, -37.14, -37.14,
         -37.14, -37.16, -37.16)
obsTime <- c(rep(1, 5), rep(26, 5))
timeUnit <- "n"
timeID <- c(rep(1, 5), rep(26, 5))
membership <- c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2)

ignition_point(lon, lat, obsTime, timeUnit, timeID, membership, "mean")

}
