% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_clustering.R
\name{local_clustering}
\alias{local_clustering}
\title{Cluster hotspots spatially}
\usage{
local_clustering(lon, lat, adjDist)
}
\arguments{
\item{lon}{numeric; a vector of longitude values.}

\item{lat}{numeric; a vector of latitude values.}

\item{adjDist}{numeric (>0); distance tolerance.
see also \code{\link{hotspot_cluster}}.}
}
\value{
integer; a vector of membership labels
}
\description{
Cluster hotspots spatially.
}
\examples{
lon <- c(141.1, 141.14, 141.12, 141.14, 141.16, 141.12, 141.14,
          141.16, 141.12, 141.14)
lat <- c(-37.10, -37.10, -37.12, -37.12, -37.12, -37.14, -37.14,
         -37.14, -37.16, -37.16)

local_clustering(lon, lat, 2000)

# 1 2 3 3 3 4 4 4 5 5
}
