% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.hotspotcluster}
\alias{plot.hotspotcluster}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{hotspotcluster}(results, draw_ignitions = TRUE, draw_hotspots = TRUE, bottom = NULL, ...)
}
\arguments{
\item{results}{a \code{hotspotcluster} object, result of \code{\link{hotspot_cluster}}.}

\item{draw_ignitions}{logical; whether or not to plot the ignitions.}

\item{draw_hotspots}{logical; whether or not to plot the hotspots.}

\item{bottom}{a \code{ggplot} object (optional). plot onto this object.}

\item{...}{further arguments will be omitted.}
}
\value{
a \code{ggplot} object.
}
\description{
\code{plot} method for class "\code{hotspotcluster}"
}
\examples{
results <- hotspot_cluster(hotspots5000,
                           lon = "lon",
                           lat = "lat",
                           obstime = "obstime",
                           ActiveTime = 24,
                           AdjDist = 3000,
                           ignition_center = "mean",
                           time_unit = "hours",
                           timestep = 1)
plot(results)
}
