% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.hotspotcluster}
\alias{plot.hotspotcluster}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{hotspotcluster}(results, hotspot = TRUE, ignition = FALSE, ...)
}
\arguments{
\item{results}{code{hotspotcluster} object, result of \code{\link{hotspot_cluster}}.}

\item{hotspot}{whether or not to plot the hotspots.}

\item{ignition}{whether or not to plot the ignitions.}
}
\value{
a \code{ggplot} object
}
\description{
\code{plot} method for class "\code{hotspotcluster}"
}
\examples{
results <- hotspot_cluster(hotspots5000,
                           lon = "lon",
                           lat = "lat",
                           obstime = "obstime",
                           ActiveTime = 24,
                           AdjDist = 3000,
                           ignition_center = "mean",
                           time_unit = "hours",
                           timestep = 1)
plot(results)
}
