% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.spotoroo}
\alias{plot.spotoroo}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{spotoroo}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{spotoroo}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{\dots}{additional arguments pass to \code{plot_spotoroo()}}
}
\description{
\code{plot} method for class "\code{spotoroo}", a simple wrapper of
\code{plot_spotoroo()}, see also \code{\link{plot_spotoroo}}.
}
\examples{
data("hotspots500")
results <- hotspot_cluster(hotspots500,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)

plot(results)
}
