% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_cluster.R
\name{plot.spotoroo}
\alias{plot.spotoroo}
\title{Plotting spatiotemporal clustering results}
\usage{
\method{plot}{spotoroo}(
  results,
  drawIgnitions = TRUE,
  drawHotspots = TRUE,
  drawNoises = TRUE,
  bottom = NULL,
  ...
)
}
\arguments{
\item{results}{an object of class "\code{spotoroo}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{drawIgnitions}{logical; if \code{TRUE}, plot the ignition points.}

\item{drawHotspots}{logical; if \code{TRUE}, plot the hotspots.}

\item{bottom}{an object of class "\code{ggplot}", optional; if \code{TRUE},
plot onto this object.}

\item{...}{further arguments will be omitted.}
}
\value{
an object of class "\code{ggplot}".
}
\description{
\code{plot} method for class "\code{spotoroo}"
}
\examples{
results <- hotspot_cluster(hotspots5000,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)
plot(results)
}
