% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_def.R
\name{plot_def}
\alias{plot_def}
\title{Default method of plotting spotoroo result}
\usage{
plot_def(
  result,
  cluster = "all",
  hotspot = TRUE,
  noise = FALSE,
  ignition = TRUE,
  from = NULL,
  to = NULL,
  bg = NULL
)
}
\arguments{
\item{result}{spotoroo result object; get from \code{\link{hotspot_cluster}}.}

\item{cluster}{integer/character; the membership labels of fires to display.
"all" stands for plotting all clusters.}

\item{hotspot}{logical; display hotspots?}

\item{noise}{logical; display noise?}

\item{ignition}{logical; display ignition points?}

\item{from}{date/datetime; only display data after this date/datetime.}

\item{to}{date/datetime; only display data before this date/datetime.}

\item{bg}{ggplot object; background image.}
}
\value{
a ggplot object
}
\description{
Default method of plotting spotoroo result. A scatter plot will be made.
}
\examples{
result <- hotspot_cluster(hotspots_fin,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

plot_def(result, cluster = 1:6, bg = plot_vic_map())
plot_def(result, cluster = "all", bg = plot_vic_map())

}
