% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fire_mov.R
\name{plot_fire_mov}
\alias{plot_fire_mov}
\title{Plotting the fire movement}
\usage{
plot_fire_mov(
  result,
  cluster = "all",
  hotspot = TRUE,
  from = NULL,
  to = NULL,
  step = 1,
  bg = NULL
)
}
\arguments{
\item{result}{\code{spotoroo} object; a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{cluster}{character/integer; if "all", plot all clusters. if a integer
vector is given, plot corresponding clusters.}

\item{hotspot}{logical; if \code{TRUE}, plot the hot spots.}

\item{from}{\strong{OPTIONAL}; date/datetime/numeric; start time; the data type
needs to be the same as the provided observed time.}

\item{to}{\strong{OPTIONAL}; date/datetime/numeric; end time; the data type
needs to be the same as the provided observed time.}

\item{step}{integer (>0); step size used in the calculation of the
fire movement.}

\item{bg}{\strong{OPTIONAL}; \code{ggplot} object; if specified, plot onto this object.}
}
\value{
A \code{ggplot} object; the plot of the fire movements.
}
\description{
This function plots the fire movement. The fire movement is calculated
from \code{\link[=get_fire_mov]{get_fire_mov()}}.
}
\examples{
if (FALSE) {
  # get clustering results
  result <- hotspot_cluster(hotspots,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

  # plot cluster 1 to 4
  plot_fire_mov(result, cluster = 1:4)

  # plot cluster 1 to 4, set step = 6
  plot_fire_mov(result, cluster = 1:4, step = 6)
}

}
