% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spotoroo.R
\name{plot_spotoroo}
\alias{plot_spotoroo}
\title{Plotting spatiotemporal clustering result}
\usage{
plot_spotoroo(
  result,
  type = "def",
  cluster = "all",
  hotspot = TRUE,
  noise = FALSE,
  ignition = TRUE,
  from = NULL,
  to = NULL,
  step = 1,
  mainBreak = NULL,
  minorBreak = NULL,
  dateLabel = NULL,
  bg = NULL
)
}
\arguments{
\item{result}{\code{spotoroo} object; a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{type}{character; type of the plot; one of "def" (default),
"timeline" (timeline) and "mov" (fire movement).}

\item{cluster}{character/integer; if "all", plot all clusters. if a integer
vector is given, plot corresponding clusters; unavailable in
\code{\link[=plot_timeline]{plot_timeline()}}.}

\item{hotspot}{logical; if \code{TRUE}, plot the hot spots; unavailable in
\code{\link[=plot_timeline]{plot_timeline()}}.}

\item{noise}{logical; if \code{TRUE}, plot the noise; only used in
\code{\link[=plot_def]{plot_def()}}.}

\item{ignition}{logical; if \code{TRUE}, plot the ignition points; only used in
\code{\link[=plot_def]{plot_def()}}.}

\item{from}{\strong{OPTIONAL}; date/datetime/numeric; start time; the data type
needs to be the same as the provided observed time.}

\item{to}{\strong{OPTIONAL}; date/datetime/numeric; end time; the data type
needs to be the same as the provided observed time.}

\item{step}{integer (>=0); step size used in the calculation of the
fire movement; only used in \code{\link[=plot_fire_mov]{plot_fire_mov()}}.}

\item{mainBreak}{\strong{OPTIONAL}; character/numeric; a string/value giving the
difference between major breaks; if the
observed time is in date/datetime
format,
this value will pass to
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}} or
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}} as
\code{date_breaks}; only used in
\code{\link[=plot_timeline]{plot_timeline()}}.}

\item{minorBreak}{\strong{OPTIONAL}; character/numeric; a string/value giving the
difference between minor breaks; if the
observed time is in date/datetime
format,
this value will pass to
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}} or
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}} as
\code{date_breaks}. only used in
\code{\link[=plot_timeline]{plot_timeline()}}.}

\item{dateLabel}{\strong{OPTIONAL}; character; a string giving the formatting
specification for the labels; if the
observed
time is in date/datetime format,
this value will pass to
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}} or
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}} as
\code{date_labels}; unavailable if the observed
time is in numeric format; only used in
\code{\link[=plot_timeline]{plot_timeline()}}.}

\item{bg}{\strong{OPTIONAL}; \code{ggplot} object; if specified, plot onto this object;
unavailable in \code{\link[=plot_timeline]{plot_timeline()}}.}
}
\value{
\code{ggplot} object; the plot of the clustering results.
}
\description{
This function takes a \code{spotoroo} object to produce a plot of the
clustering results. It can be called by \code{\link[=plot.spotoroo]{plot.spotoroo()}}.
}
\details{
if \code{type} is "def", the clustering results will be plotted spatially.
See also \code{\link[=plot_def]{plot_def()}}. Available arguments:
\itemize{
\item \code{result}
\item \code{type}
\item \code{cluster}
\item \code{ignition}
\item \code{hotspot}
\item \code{noise}
\item \code{from} (\strong{OPTIONAL})
\item \code{to} (\strong{OPTIONAL})
\item \code{bg} (\strong{OPTIONAL})
}

if \code{type} is "mov", plot of the fire movement will be made.
See also \code{\link[=plot_fire_mov]{plot_fire_mov()}}. Available arguments:
\itemize{
\item \code{result}
\item \code{type}
\item \code{cluster}
\item \code{hotspot}
\item \code{from} (\strong{OPTIONAL})
\item \code{to} (\strong{OPTIONAL})
\item \code{step}
\item \code{bg} (\strong{OPTIONAL})
}

if \code{type} is "timeline", plot of the timeline will be made.
See also \code{\link[=plot_timeline]{plot_timeline()}}. Available arguments:
\itemize{
\item \code{result}
\item \code{type}
\item \code{from} (\strong{OPTIONAL})
\item \code{to} (\strong{OPTIONAL})
\item \code{mainBreak} (\strong{OPTIONAL})
\item \code{minorBreak} (\strong{OPTIONAL})
\item \code{dateLabel} (\strong{OPTIONAL})
}
}
\examples{
if (FALSE) {
  # get clustering result
  result <- hotspot_cluster(hotspots,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          minTime = 3,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

  # different types of plots

  # default plot
  plot_spotoroo(result, "def", bg = plot_vic_map())


  # fire movement plot
  plot_spotoroo(result, "mov", cluster = 1:3, step = 3,
                bg = plot_vic_map())
}


}
