% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spotoroo.R
\name{plot_spotoroo}
\alias{plot_spotoroo}
\title{Plotting spatiotemporal clustering result}
\usage{
plot_spotoroo(
  result,
  type = "def",
  cluster = "all",
  hotspot = TRUE,
  noise = FALSE,
  ignition = TRUE,
  from = NULL,
  to = NULL,
  bg = NULL
)
}
\arguments{
\item{result}{an object of class "\code{spotoroo}",
a result of a call to \code{\link{hotspot_cluster}}.}

\item{type}{character; plot type; one of "def" and "mov".}

\item{cluster}{character/numeric; if "all", plot all clusters. if a numeric
vector is given, plot corresponding clusters.}

\item{hotspot}{logical; if \code{TRUE}, plot the hotspots.}

\item{noise}{logical; if \code{TRUE}, plot the noises.}

\item{ignition}{logical; if \code{TRUE}, plot the ignition points.}

\item{from}{time/numeric; start time, data type depends on type of observed
time.}

\item{to}{time/numeric; end time, data type depends on type of observed
time.}

\item{bg}{an object of class "\code{ggplot}", optional; if \code{TRUE},
plot onto this object. Now it only supports object without colour related
aesthetics.}
}
\value{
an object of class "\code{ggplot}".
}
\description{
can be called by \code{plot.spotoroo()}.
}
\details{
if \code{type} = "\code{def}", result will be plotted spatially,
see also \code{\link{plot_fire}}.
if \code{type} = "\code{mov}", fire movement will be plotted.
see also \code{plot_fire_mov}.
}
\examples{
data("hotspots500")
result <- hotspot_cluster(hotspots500,
                          lon = "lon",
                          lat = "lat",
                          obsTime = "obsTime",
                          activeTime = 24,
                          adjDist = 3000,
                          minPts = 4,
                          ignitionCenter = "mean",
                          timeUnit = "h",
                          timeStep = 1)

plot_spotoroo(result)
}
