% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_spotoroo.R
\name{summary_spotoroo}
\alias{summary_spotoroo}
\title{Summarizing spatiotemporal clustering results}
\usage{
summary_spotoroo(result, cluster = "all")
}
\arguments{
\item{result}{\code{spotoroo} object;
a result of a call to \code{\link[=hotspot_cluster]{hotspot_cluster()}}.}

\item{cluster}{character/integer; if "all", summarize all clusters.
if a integer vector is given, summarize corresponding
clusters.}
}
\description{
\code{summary_spotoroo()} takes a \code{spotoroo} object to produce a summary of the
clustering results. It can be called by \code{\link[=summary.spotoroo]{summary.spotoroo()}}.
}
\examples{
result <- hotspot_cluster(hotspots_fin,
                           lon = "lon",
                           lat = "lat",
                           obsTime = "obsTime",
                           activeTime = 24,
                           adjDist = 3000,
                           minPts = 4,
                           minTime = 3,
                           ignitionCenter = "mean",
                           timeUnit = "h",
                           timeStep = 1)



summary_spotoroo(result)


summary_spotoroo(result, 1:3)


}
